Added more substantial github demo
Some interesting learnings from this. Things seem to work well. It could
be nicer to chain actions. The AsyncCmd is going to end up nesting and
being verbose, but it does work and it is safe. Also, ran into the
realization that everything in the ModelValue is marked as any. So, past
the first level of the model you can have some any values. I tried to
fix it but typescript doesn't allow for nested union types so that's
that. I guess it isn't that bad. You CAN still use ImmutableModels. It
will just have to be encouraged until I find a better solution.