Bumped for publish.
Adding the re-export back in so you can just depend on this directly. I
didn't want to do it before because I was thinking of having
applications that share update/model between platforms but its better to
be able to depend on one package half the time than none of the time.